Change the block's restart point from `uint32_t` to `uint64_t`. Add a sepperate file to test issue#718. The test does not use compression because snappy does not support size bigger than `uint32_t`.

+++ table/block.h
@@ -35,7 +35,7 @@ class Block {
 
   const char* data_;
   size_t size_;
-  uint32_t restart_offset_;  // Offset in data_ of restart array
+  uint64_t restart_offset_;  // Offset in data_ of restart array
   bool owned_;               // Block owns data_[]
 };
